remove custom layout render implementation
the render method of layout has not been revisited in a long time. the layout looks like it had a custom server vs browser implementation to render because it used to manage metatags and page title. now that that responsibility has been moved to renderers, the layout can behave like a normal view.